Day 15
I tried moving it to rust to make sure my python set stuff from part 1
wasn't the issue. I had to think for a couple days to get the idea to
skip ahead to the end of the range, and at that point I figured I may as
well do it with rust.